Revert "memlat: Optimize perf event reads when possible"
This reverts commit 1d647604a0baf169928deebffe0450132c27f24b.

There are a number of problems with this optimization attempt, such as how
a perf error code could be returned to one of the counter values, a CPU
could enter an SCM call for just one event and result in garbage
computations or a divide-by-zero when calculating the stall percentage,
and IRQs are disabled as a heavy-handed way of preventing CPU migration.

Toss this commit out so that it can be replaced by something better.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Forest <forestd.github@gmail.com>